Move conformance dsl unit tests to conformance dsl module; add unit test for e-exp invocation transcription
